refactor(galeria): tornando rota get galeria publica
