+++ api/src/users/gerente/gerente.controller.ts
@@ -13,8 +13,7 @@ import { AlteraEnderecoDTO } from './dto/alteraEndereco.dto';
 import { DeletaEnderecoDTO } from './dto/deletaEndereco.dto';
 import { ContatoDTO } from './dto/contato.dto';
 import { CriarEventoDTO } from './dto/criarEvento.dto';
-import { Request, Response } from 'express';
-import { queryObjects } from 'v8';
+import { Response } from 'express';
 
 @Controller("gerente")
 export class GerenteController {
@@ -169,16 +168,6 @@ export class GerenteController {
     async BuscaTodasFotosDaGaleria(@Req() req: any, @Res() res: Response) {
         if(this.csrf.validateToken(req.cookies['x-csrf-token'] || req.headers['x-csrf-token'])){
             res.status(HttpStatus.OK).json(await this.gerenteService.buscaGaleiraPorEstabelecimento(req.user.sub))
-        }
-    }
-
-    //rota para buscar uma foto da galeira
-    @ApiOperation({summary: 'Busca por uma foto da galeira'})
-    @UseGuards(RefreshGuard)
-    @Get('/gallery/:name')
-    async BuscaFoto(@Param('name') name: string, @Req() req: any, @Res() res: Response) {
-        if(this.csrf.validateToken(req.cookies['x-csrf-token'] || req.headers['x-csrf-token'])) {
-            res.status(HttpStatus.OK).sendFile(await this.gerenteService.buscaFotoGaleria(req.user.sub, name))
         }
     }
